+++ R/GLAM_2d_covariates.R
@@ -211,6 +211,7 @@ GLAM_2d_covariates <- function(R, Y,
   SE <- sqrt(diag(Cov))
   Cov_Alpha <- Cov[1:(cu * cs), 1:(cu * cs)]
   Cov_beta <- Cov[(cu * cs + 1):(cu * cs + p), (cu * cs + 1):(cu * cs + p)]
+  Cov_Alpha_beta <- Cov[1:(cu * cs), (cu * cs + 1):(cu * cs + p)]
 
   # ---- Save results in list ----
   Alpha <- matrix(Alpha, nrow = cu, ncol = cs)
@@ -228,6 +229,7 @@ GLAM_2d_covariates <- function(R, Y,
     beta = beta,
     Cov_beta = Cov_beta,
     SE_beta = SE_beta,
+    Cov_Alpha_beta = Cov_Alpha_beta,
     Eta0 = Base,
     H = H,
     deviance = dev,